[Module JS] Utilisation des modules et création de classe externe au fichier script.js
