+++ 02-snake/index.html
@@ -2,7 +2,7 @@
 <html>
     <head>
         <title>Jeu du Serpent</title>
-        <script src="script.js"></script>
+        <script type="module" src="./script.js"></script>
     </head>
     <body>
     </body>
